Extract TeamOverviewChartTooltip and TeamOverviewUnansweredIssuesChart into separate component files
